requirements.txt
streamlit
openai
python-dotenv
reportlab
PyPDF2 
python-docx
pandas 
openpyxl